initializing current live mark to NULL

+++ runtime/gc/accounting/atomic_stack.h
@@ -214,7 +214,7 @@ class StructuredAtomicStack {
     DCHECK_LT(static_cast<size_t>(index), stack_data_->capacity_);
     //stack_data_->back_index_ = index + 1;
 
-    if(index >= stack_data_->capacity_) {
+    if(static_cast<size_t>(index) >= stack_data_->capacity_) {
       LOG(FATAL) << "ERROR in pushing back " << index << "; capacity = " <<
           stack_data_->capacity_;
     }
